2025/03/21
・実家で作業するからといり合えずプッシュ
・User系処理のリファクタリング中

+++ src/RaremintShop.WebHost/Controllers/AdminController.cs
@@ -2,38 +2,94 @@
 using Microsoft.AspNetCore.Mvc;
 using RaremintShop.Module.Identity.Models;
 using RaremintShop.Module.Identity.Services;
+using static RaremintShop.Shared.Constants;
 
 namespace RaremintShop.WebHost.Controllers
 {
-    [Authorize(Roles = "Admin")] // 管理者のみアクセス可
+    [Authorize(Roles = Roles.Admin)] // 管理者のみアクセス可
     public class AdminController : Controller
     {
+        /// <summary>ユーザー管理、認証、およびロール管理のためのサービス</summary>
         private readonly IUserService _userService;
 
-        public AdminController(IUserService userService)
+        /// <summary>ロガー</summary>
+        private readonly ILogger<AdminController> _logger;
+
+        /// <summary>
+        /// AdminController クラスの新しいインスタンスを初期化します。
+        /// </summary>
+        /// <param name="userService">UserServiceのインターフェース</param>
+        /// <param name="logger">ロギングのためのILogger</param>
+        public AdminController(IUserService userService, ILogger<AdminController> logger)
         {
             _userService = userService;
+            _logger = logger;
         }
 
+
+        /// <summary>
+        /// 管理者ダッシュボードページの表示
+        /// </summary>
+        /// <returns>管理者ダッシュボードページ</returns>
         public IActionResult DashBoard()
         {
             return View();
         }
 
+        /// <summary>
+        /// ユーザー管理ページの表示
+        /// </summary>
+        /// <returns>ユーザー管理ページ</returns>
         [HttpGet]
         public async Task<IActionResult> UserManagement()
         {
-            var users = await _userService.GetAllUsersAsync();
-            return View(users);
+            try
+            {
+                var users = await _userService.GetAllUsersAsync();
+                return View(users);
+            }
+            catch (Exception ex)
+            {
+                // エラーメッセージを表示
+                _logger.LogError(ex, "ユーザー情報の取得中にエラーが発生しました。");
+                ModelState.AddModelError(string.Empty, "ユーザー情報の取得中にエラーが発生しました。");
+                return View(new List<UserManagementViewModel>());
+            }
         }
 
+        /// <summary>
+        /// ユーザー編集ページの表示
+        /// </summary>
+        /// <param name="id">ユーザーID</param>
+        /// <returns>ユーザー編集ページ</returns>
         [HttpGet]
         public async Task<IActionResult> Edit(string id)
         {
-            var user = await _userService.GetByIdForEditAsync(id);
-            return View(user);
+            try
+            {
+                var user = await _userService.GetByIdForEditAsync(id);
+                if (user == null)
+                {
+                    return NotFound();
+                }
+                return View(user);
+            }
+            catch (Exception ex)
+            {
+                // エラーメッセージを表示
+                _logger.LogError(ex, "ユーザー情報の取得中にエラーが発生しました。");
+                ModelState.AddModelError(string.Empty, "ユーザー情報の取得中にエラーが発生しました。");
+                return RedirectToAction(RedirectPaths.AdminUserManagement, RedirectPaths.AdminController);
+            }
         }
 
+
+
+        /// <summary>
+        /// ユーザー編集フォームからのPOST処理
+        /// </summary>
+        /// <param name="model">ユーザー編集のためのモデル</param>
+        /// <returns>編集結果に応じたビュー</returns>
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(UserEditViewModel model)
@@ -41,26 +97,64 @@ namespace RaremintShop.WebHost.Controllers
             if (!ModelState.IsValid)
             {
                 var roles = await _userService.GetAllRolesAsync();
-                model.AvailableRoles = roles.Select(x => x.Name).ToList();
+                model.AvailableRoles = roles.Select(x => x.Name!).ToList();
+                return View(model);
             }
-            else
+
+            try
             {
                 var result = await _userService.UpdateUserAsync(model);
                 if (result.Succeeded)
                 {
-                    return RedirectToAction("UserManagement");
+                    return RedirectToAction(RedirectPaths.AdminUserManagement, RedirectPaths.AdminController);
+                }
+                else
+                {
+                    // エラーメッセージを表示
+                    ModelState.AddModelError(string.Empty, "ユーザー情報の更新中にエラーが発生しました。");
+                    return View(model);
                 }
             }
-            return View(model);
+            catch (Exception ex)
+            {
+                // ログの記録
+                _logger.LogError(ex, "ユーザー情報の更新中にエラーが発生しました。");
+
+                // エラーメッセージの表示
+                ModelState.AddModelError(string.Empty, "予期しないエラーが発生しました。");
+                return View(model);
+            }
         }
 
+
+        /// <summary>
+        /// ユーザー削除処理
+        /// </summary>
+        /// <param name="id">ユーザーID</param>
+        /// <returns>ユーザー管理ページにリダイレクト</returns>
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Delete(string id)
         {
-            var user = await _userService.GetByIdAsync(id);
-            await _userService.DeleteUserAsync(user);
-            return RedirectToAction("UserManagement");
+            try
+            {
+                var user = await _userService.GetByIdAsync(id);
+                if (user == null)
+                {
+                    return NotFound();
+                }
+                await _userService.DeleteUserAsync(user);
+                return RedirectToAction(RedirectPaths.AdminUserManagement, RedirectPaths.AdminController);
+            }
+            catch (Exception ex)
+            {
+                // ログの記録
+                _logger.LogError(ex, "ユーザー削除中にエラーが発生しました。");
+
+                // エラーメッセージの表示
+                ModelState.AddModelError(string.Empty, "ユーザー削除中にエラーが発生しました。");
+                return RedirectToAction("UserManagement");
+            }
         }
     }
 }
